removed unnecessary casts

+++ RomanNumerals/src/main/java/edu/pdx/cs410J/ahsiao/Kata.java
@@ -13,6 +13,7 @@ public class Kata {
     System.err.println("Missing command line arguments");
     System.exit(1);
   }
+
   public static String decToRoman(int decimal) {
     if(decimal > 3000){
       return "Too large";
@@ -20,57 +21,57 @@ public class Kata {
 
     String numeral = "";
     if (decimal >= 1000) {
-      for (int i = 0; i < (int)(decimal / 1000); i++) {
+      for (int i = 0; i < (decimal / 1000); i++) {
         numeral += "M";
       }
-      decimal -= (int)(decimal / 1000) * 1000;
+      decimal -= (decimal / 1000) * 1000;
     }
 
     if(decimal >= 900){
       numeral += "CM";
       decimal -= 900;
     }else if(decimal >= 500) {
-      for (int i = 0; i < (int)(decimal / 500); i++) {
+      for (int i = 0; i < (decimal / 500); i++) {
         numeral += "D";
       }
-      decimal -= (int)(decimal / 500) * 500;
+      decimal -= (decimal / 500) * 500;
     }
 
     if(decimal >= 400){
       numeral += "CD";
       decimal -= 400;
     } else if(decimal >= 100) {
-      for (int i = 0; i < (int)(decimal / 100); i++) {
+      for (int i = 0; i < (decimal / 100); i++) {
         numeral += "C";
       }
-      decimal -= (int)(decimal / 100) * 100;
+      decimal -= (decimal / 100) * 100;
     }
 
     if(decimal >= 90){
       numeral += "XC";
       decimal -= 90;
     }else if(decimal >= 50) {
-      for (int i = 0; i < (int)(decimal / 50); i++) {
+      for (int i = 0; i < (decimal / 50); i++) {
         numeral += "L";
       }
-      decimal -= (int)(decimal / 50) * 50;
+      decimal -= (decimal / 50) * 50;
     }
 
     if(decimal >= 40){
       numeral += "XL";
       decimal -= 40;
     }else if(decimal >= 10) {
-      for (int i = 0; i < (int)(decimal / 10); i++) {
+      for (int i = 0; i < (decimal / 10); i++) {
         numeral += "X";
       }
-      decimal -= (int)(decimal / 10) * 10;
+      decimal -= (decimal / 10) * 10;
     }
 
     if(decimal >= 9){
       numeral += "IX";
       decimal -= 9;
     } else if(decimal >= 5) {
-      for (int i = 0; i < (int)(decimal / 5); i++) {
+      for (int i = 0; i < (decimal / 5); i++) {
         numeral += "V";
       }
       decimal -= (decimal / 5) * 5;
